pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = ""
authors = [
    {name = "Ben Lawrence",email = "benlawrencenz@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-browser-reload (>=1.18.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
